rake: use built-in rspec functions
In the conversion to bundler, a09dcbdced6418e32ec69466eb90de6355e83937,
the default rake task was changed to simply shell out to rspec. This
commit restores the change to the Rakefile so that we continue to use
the built-in rspec rake task, along with the --backtrace and --color
customizations.