Fixes #140 (travis testing for other ruby impls)
This adds JRuby in 1.9 mode (which should track the 1.7 branch) and also JRuby 9000 which was recently released (it's ruby 2.2.2). These are run on three different JVMs as well. It should also test against Rubinus 2.

There's also no point in testing on the different jvms for straight-up mri so those combinations are excluded.