Fix NoMethodError bug due to nil worker
We replace Parallel::Worker object in workers
by `Array#[]=` in the isolation mode.
The assignment generates `nil` element in workers
when `worker.size` is below the specified index.

In the previous code, this `nil` element
resulted in NoMethodError when `worker.each` was called.
This patch fixes the bug by calling `Array#compact` beforehand.